LPS-180450 Better organization of the code
