+++ modules/util/portal-tools-db-virtual-instance-migration/src/main/java/com/liferay/portal/tools/db/virtual/instance/migration/VirtualInstanceMigration.java
@@ -16,16 +16,12 @@ package com.liferay.portal.tools.db.virtual.instance.migration;
 
 import com.liferay.portal.tools.db.virtual.instance.migration.error.ErrorCodes;
 import com.liferay.portal.tools.db.virtual.instance.migration.internal.util.Database;
-import com.liferay.portal.tools.db.virtual.instance.migration.internal.util.Version;
+import com.liferay.portal.tools.db.virtual.instance.migration.internal.validation.Validator;
+import com.liferay.portal.tools.db.virtual.instance.migration.internal.validation.ValidatorRecorder;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -109,7 +105,7 @@ public class VirtualInstanceMigration {
 
 				if (!Database.isSingleVirtualInstance(_sourceConnection)) {
 					System.err.println(
-						"ERROR: Source database has several instances, that " +
+						"ERROR: Source database has several instances. That " +
 							"is not supported by the tool");
 
 					_exitWithCode(ErrorCodes.SOURCE_MULTI_INSTANCES);
@@ -123,9 +119,14 @@ public class VirtualInstanceMigration {
 					_exitWithCode(ErrorCodes.DESTINATION_NOT_DEFAULT);
 				}
 
-				if (!_validate(_sourceConnection, _destinationConnection)) {
+				if (!Validator.validateDatabases(
+						_sourceConnection, _destinationConnection)) {
+
+					ValidatorRecorder.printMessages();
 					_exitWithCode(ErrorCodes.VALIDATION_ERROR);
 				}
+
+				System.out.println("All validations passed successfully");
 			}
 			catch (ParseException parseException) {
 				System.err.println(
@@ -149,68 +150,6 @@ public class VirtualInstanceMigration {
 			exception.printStackTrace();
 			_exitWithCode(ErrorCodes.UNEXPECTED_ERROR);
 		}
-	}
-
-	private static boolean _checkTables(
-			Connection sourceConnection, Connection destinationConnection)
-		throws SQLException {
-
-		boolean valid = true;
-
-		List<String> sourceTables = Database.getTables(sourceConnection);
-		List<String> destinationTables = Database.getTables(
-			destinationConnection);
-
-		for (String tableName : sourceTables) {
-			if (destinationTables.contains(tableName)) {
-				destinationTables.remove(tableName);
-
-				continue;
-			}
-
-			System.out.println(
-				"WARNING: Table " + tableName +
-					" is not present in destination database");
-			valid = false;
-		}
-
-		if (!destinationTables.isEmpty()) {
-			for (String tableName : destinationTables) {
-				System.out.println(
-					"WARNING: Table " + tableName +
-						" is not present in source database");
-			}
-
-			valid = false;
-		}
-
-		return valid;
-	}
-
-	private static boolean _checkWebIds(
-			Connection sourceConnection, Connection destinationConnection)
-		throws SQLException {
-
-		String sourceWebId = Database.getWebId(sourceConnection);
-
-		try (PreparedStatement preparedStatement =
-				destinationConnection.prepareStatement(
-					"select companyId from Company where webId = ?")) {
-
-			preparedStatement.setString(1, sourceWebId);
-
-			try (ResultSet resultSet = preparedStatement.executeQuery()) {
-				if (resultSet.next()) {
-					System.out.println(
-						"WARNING: webId " + sourceWebId +
-							" already exists in destination database");
-
-					return false;
-				}
-			}
-		}
-
-		return true;
 	}
 
 	private static void _exitWithCode(int code) {
@@ -259,175 +198,6 @@ public class VirtualInstanceMigration {
 			"su", "source-user", true, "Set the source database user name.");
 
 		return options;
-	}
-
-	private static void _printErrorMessages(
-		List<String> modules, String message) {
-
-		for (String module : modules) {
-			System.out.println("ERROR: Module " + module + message);
-		}
-	}
-
-	private static void _printWarningMessages(
-		List<String> modules, String message) {
-
-		for (String module : modules) {
-			System.out.println("WARNING: Module " + module + message);
-		}
-	}
-
-	private static boolean _validate(
-			Connection sourceConnection, Connection destinationConnection)
-		throws SQLException {
-
-		boolean valid = true;
-
-		if (!_validateReleaseTableState(sourceConnection)) {
-			System.out.println(
-				"ERROR: Source database Release_ table has records with an " +
-					"invalid state_");
-			valid = false;
-		}
-
-		if (!_validateReleaseTableState(destinationConnection)) {
-			System.out.println(
-				"ERROR: Destination database Release_ table has records with " +
-					"an invalid state_");
-			valid = false;
-		}
-
-		valid &= _validateReleaseTableModules(
-			sourceConnection, destinationConnection);
-
-		valid &= _checkTables(sourceConnection, destinationConnection);
-
-		valid &= _checkWebIds(sourceConnection, destinationConnection);
-
-		return valid;
-	}
-
-	private static boolean _validateReleaseTableModules(
-			Connection sourceConnection, Connection destinationConnection)
-		throws SQLException {
-
-		boolean valid = true;
-
-		try (PreparedStatement preparedStatement1 =
-				sourceConnection.prepareStatement(
-					"select servletContextName, schemaVersion, verified from " +
-						" Release_");
-			ResultSet resultSet1 = preparedStatement1.executeQuery()) {
-
-			List<String> missingModules = new ArrayList<>();
-			List<String> missingServiceModules = new ArrayList<>();
-			List<String> lowerVersionModules = new ArrayList<>();
-			List<String> higherVersionModules = new ArrayList<>();
-			List<String> sourceUnverifiedModules = new ArrayList<>();
-			List<String> destinationUnverifiedModules = new ArrayList<>();
-
-			while (resultSet1.next()) {
-				String sourceServletContextName = resultSet1.getString(1);
-				Version sourceVersion = Version.parseVersion(
-					resultSet1.getString(2));
-				boolean sourceVerified = resultSet1.getBoolean(3);
-
-				try (PreparedStatement preparedStatement2 =
-						destinationConnection.prepareStatement(
-							"select servletContextName, schemaVersion, " +
-								"verified from Release_ where " +
-									"servletContextName = ?")) {
-
-					preparedStatement2.setString(1, sourceServletContextName);
-
-					try (ResultSet resultSet2 =
-							preparedStatement2.executeQuery()) {
-
-						if (!resultSet2.next()) {
-							if (sourceServletContextName.endsWith(".service")) {
-								missingServiceModules.add(
-									sourceServletContextName);
-							}
-							else {
-								missingModules.add(sourceServletContextName);
-							}
-
-							continue;
-						}
-
-						Version destinationVersion = Version.parseVersion(
-							resultSet2.getString(2));
-						boolean destinationVerified = resultSet2.getBoolean(3);
-
-						if (sourceVersion.compareTo(destinationVersion) < 0) {
-							lowerVersionModules.add(sourceServletContextName);
-						}
-						else if (sourceVersion.compareTo(destinationVersion) >
-									0) {
-
-							higherVersionModules.add(sourceServletContextName);
-						}
-
-						if (sourceVerified && !destinationVerified) {
-							destinationUnverifiedModules.add(
-								sourceServletContextName);
-						}
-						else if (!sourceVerified && destinationVerified) {
-							sourceUnverifiedModules.add(
-								sourceServletContextName);
-						}
-					}
-				}
-			}
-
-			_printErrorMessages(
-				missingServiceModules,
-				" will not be available in the destination");
-			_printErrorMessages(
-				lowerVersionModules,
-				" needs to be upgraded in source database before the " +
-					"migration");
-			_printErrorMessages(
-				higherVersionModules,
-				" is in a lower version in destination database");
-			_printErrorMessages(
-				sourceUnverifiedModules,
-				" needs to be verified in the source before the migration");
-			_printErrorMessages(
-				destinationUnverifiedModules,
-				" needs to be verified in the destination before the " +
-					"migration");
-			_printWarningMessages(
-				missingModules, " will not be available in the destination");
-
-			if (!missingModules.isEmpty() || !missingServiceModules.isEmpty() ||
-				!lowerVersionModules.isEmpty() ||
-				!higherVersionModules.isEmpty() ||
-				!sourceUnverifiedModules.isEmpty() ||
-				!destinationUnverifiedModules.isEmpty()) {
-
-				valid = false;
-			}
-		}
-
-		return valid;
-	}
-
-	private static boolean _validateReleaseTableState(Connection connection)
-		throws SQLException {
-
-		boolean stateOk = true;
-
-		try (PreparedStatement preparedStatement = connection.prepareStatement(
-				"select servletContextName from Release_ where state_ != 0;");
-			ResultSet resultSet = preparedStatement.executeQuery()) {
-
-			if (resultSet.next()) {
-				stateOk = false;
-			}
-		}
-
-		return stateOk;
 	}
 
 	private static Connection _destinationConnection;
